backend-python/requirements.txt
# src dependencies
flask
psycopg2-binary

# test dependencies
pytest